+++ src/main/java/edu/colostate/cs/cs414/cteam/p3/model/Move.java
@@ -1,0 +1,57 @@
+package edu.colostate.cs.cs414.cteam.p3.model;
+
+public class Move {
+	private int fromX;
+	private int toX;
+	private int fromY;
+	private int toY;
+	private boolean attack;
+	
+	public Move(int fromX, int fromY, int toX, int toY, boolean attack) {
+		this.fromX = fromX;
+		this.fromY = fromY;
+		this.toX = toX;
+		this.toY = toY;
+		this.attack = attack;
+	}
+
+	public int getFromX() {
+		return fromX;
+	}
+
+	public void setFromX(int fromX) {
+		this.fromX = fromX;
+	}
+
+	public int getToX() {
+		return toX;
+	}
+
+	public void setToX(int toX) {
+		this.toX = toX;
+	}
+
+	public int getFromY() {
+		return fromY;
+	}
+
+	public void setFromY(int fromY) {
+		this.fromY = fromY;
+	}
+
+	public int getToY() {
+		return toY;
+	}
+
+	public void setToY(int toY) {
+		this.toY = toY;
+	}
+
+	public boolean isAttack() {
+		return attack;
+	}
+
+	public void setAttack(boolean attack) {
+		this.attack = attack;
+	}
+}
