+++ src/main/java/edu/colostate/cs/cs414/cteam/p3/model/Tile.java
@@ -1,3 +1,4 @@
+package edu.colostate.cs.cs414.cteam.p3.model;
 
 public class Tile {
