+++ src/module-info.java
@@ -1,0 +1,9 @@
+/**
+ * 
+ */
+/**
+ * @author ian
+ *
+ */
+module jungle {
+}
